[fix] Moved base tech object classes to files
