+++ src/EasyEPlanner.csproj
@@ -190,6 +190,19 @@
     <Compile Include="TechObject\Base\BaseOperation.cs" />
     <Compile Include="TechObject\Base\BaseProperty.cs" />
     <Compile Include="TechObject\Base\BaseTechObject.cs" />
+    <Compile Include="TechObject\Base\TechObjects\BaseBoiler.cs" />
+    <Compile Include="TechObject\Base\TechObjects\BaseCooler.cs" />
+    <Compile Include="TechObject\Base\TechObjects\BaseHeater.cs" />
+    <Compile Include="TechObject\Base\TechObjects\BaseLine.cs" />
+    <Compile Include="TechObject\Base\TechObjects\BaseLineIn.cs" />
+    <Compile Include="TechObject\Base\TechObjects\BaseLineOut.cs" />
+    <Compile Include="TechObject\Base\TechObjects\BaseMaster.cs" />
+    <Compile Include="TechObject\Base\TechObjects\BaseAutomat.cs" />
+    <Compile Include="TechObject\Base\TechObjects\BaseMixer.cs" />
+    <Compile Include="TechObject\Base\TechObjects\BasePost.cs" />
+    <Compile Include="TechObject\Base\TechObjects\BasePOU.cs" />
+    <Compile Include="TechObject\Base\TechObjects\BaseTank.cs" />
+    <Compile Include="TechObject\Base\TechObjects\BaseWaterTank.cs" />
     <Compile Include="TechObject\LocalRestriction.cs" />
     <Compile Include="TechObject\Mode.cs" />
     <Compile Include="TechObject\ModesManager.cs" />
