+++ src/TechObject/Base/BaseTechObject.cs
@@ -161,277 +161,4 @@ namespace TechObject
         private BaseOperation[] objectOperations;
         private BaseProperty[] objectProperties;
     }
-
-    /// <summary>
-    /// Базовый автомат
-    /// </summary>
-    public class BaseAutomat : BaseTechObject
-    {
-        public BaseAutomat() : base()
-        {
-            S88Level = 2;
-            Name = "Автомат";
-            EplanName = "automat";
-            BaseOperations = DataBase.Imitation.BaseEmptyOperations();
-            BaseProperties = DataBase.Imitation.AutomatProperties();
-            BasicName = "automat";
-        }
-
-        public override BaseTechObject Clone()
-        {
-            return new BaseAutomat();
-        }
-    }
-
-    /// <summary>
-    /// Базовый мастер
-    /// </summary>
-    public class BaseMaster : BaseTechObject
-    {
-        public BaseMaster() : base()
-        {
-            S88Level = 1;
-            Name = "Мастер";
-            EplanName = "master";
-            BaseOperations = DataBase.Imitation.BaseEmptyOperations();
-            BaseProperties = DataBase.Imitation.EmptyProperties();
-            BasicName = "master";
-        }
-
-        public override BaseTechObject Clone()
-        {
-            return new BaseMaster();
-        }
-    }
-
-	/// <summary>
-    /// Базовый бачок
-    /// </summary>
-    public class BaseWaterTank : BaseTechObject
-    {
-        public BaseWaterTank() : base()
-        {
-            S88Level = 2;
-            Name = "Бачок";
-            EplanName = "_tank";
-            BaseOperations = DataBase.Imitation.BaseEmptyOperations();
-            BaseProperties = DataBase.Imitation.EmptyProperties();
-            BasicName = "cooler";
-        }
-
-        public override BaseTechObject Clone()
-        {
-            return new BaseWaterTank();
-        }
-    }
-
-    /// <summary>
-    /// Базовый бойлер
-    /// </summary>
-    public class BaseBoiler : BaseTechObject
-    {
-        public BaseBoiler() : base()
-        {
-            S88Level = 2;
-            Name = "Бойлер";
-            EplanName = "boil";
-            BaseOperations = DataBase.Imitation.BaseEmptyOperations();
-            BaseProperties = DataBase.Imitation.EmptyProperties();
-            BasicName = "boiler";
-        }
-
-        public override BaseTechObject Clone()
-        {
-            return new BaseBoiler();
-        }
-    }
-
-    /// <summary>
-    /// Базовая линия
-    /// </summary>
-    public class BaseLine : BaseTechObject
-    {
-        public BaseLine() : base()
-        {
-            S88Level = 2;
-            Name = "Линия";
-            EplanName = "line";
-            BaseOperations = DataBase.Imitation.BaseLineOperations();
-            BaseProperties = DataBase.Imitation.LineProperties();
-            BasicName = "line";
-        }
-
-        public override BaseTechObject Clone()
-        {
-            return new BaseLine();
-        }
-    }
-
-    /// <summary>
-    /// Базовая линия приемки
-    /// </summary>
-    public class BaseLineIn : BaseTechObject
-    {
-        public BaseLineIn() : base()
-        {
-            S88Level = 2;
-            Name = "Линия приемки";
-            EplanName = "line";
-            BaseOperations = DataBase.Imitation.BaseLineOperations();
-            BaseProperties = DataBase.Imitation.LineProperties();
-            BasicName = "line";
-        }
-
-        public override BaseTechObject Clone()
-        {
-            return new BaseLineIn();
-        }
-    }
-
-    /// <summary>
-    /// Базовая линия выдачи
-    /// </summary>
-    public class BaseLineOut : BaseTechObject
-    {
-        public BaseLineOut() : base()
-        {
-            S88Level = 2;
-            Name = "Линия выдачи";
-            EplanName = "line";
-            BaseOperations = DataBase.Imitation.BaseLineOperations();
-            BaseProperties = DataBase.Imitation.LineProperties();
-            BasicName = "line";
-        }
-
-        public override BaseTechObject Clone()
-        {
-            return new BaseLineOut();
-        }
-    }
-
-    /// <summary>
-    /// Базовый пастеризатор
-    /// </summary>
-    public class BasePOU : BaseTechObject
-    {
-        public BasePOU() : base()
-        {
-            S88Level = 2;
-            Name = "Пастеризатор";
-            EplanName = "pasteurizator";
-            BaseOperations = DataBase.Imitation.BaseEmptyOperations();
-            BaseProperties = DataBase.Imitation.EmptyProperties();
-            BasicName = "pasteurizator";
-        }
-
-        public override BaseTechObject Clone()
-        {
-            return new BasePOU();
-        }
-    }
-
-    /// <summary>
-    /// Базовый пост
-    /// </summary>
-    public class BasePost : BaseTechObject
-    {
-        public BasePost() : base()
-        {
-            S88Level = 2;
-            Name = "Пост";
-            EplanName = "post";
-            BaseOperations = DataBase.Imitation.BaseEmptyOperations();
-            BaseProperties = DataBase.Imitation.EmptyProperties();
-            BasicName = "post";
-        }
-
-        public override BaseTechObject Clone()
-        {
-            return new BasePost();
-        }
-    }
-
-    /// <summary>
-    /// Базовый танк
-    /// </summary>
-    public class BaseTank : BaseTechObject
-    {
-        public BaseTank() : base()
-        {
-            S88Level = 1;
-            Name = "Танк";
-            EplanName = "tank";
-            BaseOperations = DataBase.Imitation.BaseTankOperations();
-            BaseProperties = DataBase.Imitation.TankProperties();
-            BasicName = "tank";
-        }
-
-        public override BaseTechObject Clone()
-        {
-            return new BaseTank();
-        }
-    }
-
-    /// <summary>
-    /// Базовый узел подогрева
-    /// </summary>
-    public class BaseHeater : BaseTechObject
-    {
-        public BaseHeater() : base()
-        {
-            S88Level = 2;
-            Name = "Узел подогрева";
-            EplanName = "heater_node";
-            BaseOperations = DataBase.Imitation.BaseEmptyOperations();
-            BaseProperties = DataBase.Imitation.EmptyProperties();
-            BasicName = "heater_node";
-        }
-
-        public override BaseTechObject Clone()
-        {
-            return new BaseHeater();
-        }
-    }
-
-    /// <summary>
-    /// Базовый узел охлаждения
-    /// </summary>
-    public class BaseCooler : BaseTechObject
-    {
-        public BaseCooler() : base()
-        {
-            S88Level = 2;
-            Name = "Узел охлаждения";
-            EplanName = "cooler_node";
-            BaseOperations = DataBase.Imitation.BaseEmptyOperations();
-            BaseProperties = DataBase.Imitation.EmptyProperties();
-            BasicName = "cooler_node";
-        }
-
-        public override BaseTechObject Clone()
-        {
-            return new BaseCooler();
-        }
-    }
-
-    /// <summary>
-    /// Базовый узел перемешивания
-    /// </summary>
-    public class BaseMixer : BaseTechObject
-    {
-        public BaseMixer() : base()
-        {
-            S88Level = 2;
-            Name = "Узел перемешивания";
-            EplanName = "mix_node";
-            BaseOperations = DataBase.Imitation.BaseEmptyOperations();
-            BaseProperties = DataBase.Imitation.EmptyProperties();
-            BasicName = "mix_node";
-        }
-
-        public override BaseTechObject Clone()
-        {
-            return new BaseMixer();
-        }
-    }
 }
