+++ src/TechObject/Base/TechObjects/BaseBoiler.cs
@@ -1,0 +1,29 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TechObject
+{
+    /// <summary>
+    /// Базовый бойлер
+    /// </summary>
+    public class BaseBoiler : BaseTechObject
+    {
+        public BaseBoiler() : base()
+        {
+            S88Level = 2;
+            Name = "Бойлер";
+            EplanName = "boil";
+            BaseOperations = DataBase.Imitation.BaseEmptyOperations();
+            BaseProperties = DataBase.Imitation.EmptyProperties();
+            BasicName = "boiler";
+        }
+
+        public override BaseTechObject Clone()
+        {
+            return new BaseBoiler();
+        }
+    }
+}
