+++ src/TechObject/Base/TechObjects/BaseCooler.cs
@@ -1,0 +1,29 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TechObject
+{
+    /// <summary>
+    /// Базовый узел охлаждения
+    /// </summary>
+    public class BaseCooler : BaseTechObject
+    {
+        public BaseCooler() : base()
+        {
+            S88Level = 2;
+            Name = "Узел охлаждения";
+            EplanName = "cooler_node";
+            BaseOperations = DataBase.Imitation.BaseEmptyOperations();
+            BaseProperties = DataBase.Imitation.EmptyProperties();
+            BasicName = "cooler_node";
+        }
+
+        public override BaseTechObject Clone()
+        {
+            return new BaseCooler();
+        }
+    }
+}
