+++ src/TechObject/Base/TechObjects/BaseHeater.cs
@@ -1,0 +1,29 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TechObject
+{
+    /// <summary>
+    /// Базовый узел подогрева
+    /// </summary>
+    public class BaseHeater : BaseTechObject
+    {
+        public BaseHeater() : base()
+        {
+            S88Level = 2;
+            Name = "Узел подогрева";
+            EplanName = "heater_node";
+            BaseOperations = DataBase.Imitation.BaseEmptyOperations();
+            BaseProperties = DataBase.Imitation.EmptyProperties();
+            BasicName = "heater_node";
+        }
+
+        public override BaseTechObject Clone()
+        {
+            return new BaseHeater();
+        }
+    }
+}
