+++ src/TechObject/Base/TechObjects/BaseLineOut.cs
@@ -1,0 +1,29 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TechObject
+{
+    /// <summary>
+    /// Базовая линия выдачи
+    /// </summary>
+    public class BaseLineOut : BaseTechObject
+    {
+        public BaseLineOut() : base()
+        {
+            S88Level = 2;
+            Name = "Линия выдачи";
+            EplanName = "line";
+            BaseOperations = DataBase.Imitation.BaseLineOperations();
+            BaseProperties = DataBase.Imitation.LineProperties();
+            BasicName = "line";
+        }
+
+        public override BaseTechObject Clone()
+        {
+            return new BaseLineOut();
+        }
+    }
+}
