+++ src/TechObject/Base/TechObjects/BaseMaster.cs
@@ -1,0 +1,29 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TechObject
+{
+    /// <summary>
+    /// Базовый мастер
+    /// </summary>
+    public class BaseMaster : BaseTechObject
+    {
+        public BaseMaster() : base()
+        {
+            S88Level = 1;
+            Name = "Мастер";
+            EplanName = "master";
+            BaseOperations = DataBase.Imitation.BaseEmptyOperations();
+            BaseProperties = DataBase.Imitation.EmptyProperties();
+            BasicName = "master";
+        }
+
+        public override BaseTechObject Clone()
+        {
+            return new BaseMaster();
+        }
+    }
+}
