+++ src/TechObject/Base/TechObjects/BaseMixer.cs
@@ -1,0 +1,29 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TechObject
+{
+    /// <summary>
+    /// Базовый узел перемешивания
+    /// </summary>
+    public class BaseMixer : BaseTechObject
+    {
+        public BaseMixer() : base()
+        {
+            S88Level = 2;
+            Name = "Узел перемешивания";
+            EplanName = "mix_node";
+            BaseOperations = DataBase.Imitation.BaseEmptyOperations();
+            BaseProperties = DataBase.Imitation.EmptyProperties();
+            BasicName = "mix_node";
+        }
+
+        public override BaseTechObject Clone()
+        {
+            return new BaseMixer();
+        }
+    }
+}
