+++ src/TechObject/Base/TechObjects/BasePOU.cs
@@ -1,0 +1,29 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TechObject
+{
+    /// <summary>
+    /// Базовый пастеризатор
+    /// </summary>
+    public class BasePOU : BaseTechObject
+    {
+        public BasePOU() : base()
+        {
+            S88Level = 2;
+            Name = "Пастеризатор";
+            EplanName = "pasteurizator";
+            BaseOperations = DataBase.Imitation.BaseEmptyOperations();
+            BaseProperties = DataBase.Imitation.EmptyProperties();
+            BasicName = "pasteurizator";
+        }
+
+        public override BaseTechObject Clone()
+        {
+            return new BasePOU();
+        }
+    }
+}
